refactor and restructure Object Results
