+++ explorer/client/src/pages/object-result/ObjectResult.tsx
@@ -14,12 +14,12 @@ import { NetworkContext } from '../../context';
 import theme from '../../styles/theme.module.css';
 import { DefaultRpcClient as rpc } from '../../utils/api/DefaultRpcClient';
 import { IS_STATIC_ENV } from '../../utils/envUtil';
-import ObjectLoaded from './ObjectLoaded';
 import {
     instanceOfDataType,
     translate,
     type DataType,
 } from './ObjectResultType';
+import ObjectView from './views/ObjectView';
 
 const DATATYPE_DEFAULT: DataType = {
     id: '',
@@ -92,7 +92,7 @@ const ObjectResultAPI = ({ objID }: { objID: string }): JSX.Element => {
     }, [objID, network]);
 
     if (showObjectState.loadState === 'loaded') {
-        return <ObjectLoaded data={showObjectState as DataType} />;
+        return <ObjectView data={showObjectState as DataType} />;
     }
     if (showObjectState.loadState === 'pending') {
         return (
@@ -111,10 +111,10 @@ const ObjectResultStatic = ({ objID }: { objID: string }): JSX.Element => {
     const data = findDataFromID(objID, undefined);
 
     if (instanceOfDataType(data)) {
-        return <ObjectLoaded data={data} />;
+        return <ObjectView data={data} />;
     } else {
         try {
-            return <ObjectLoaded data={translate(data)} />;
+            return <ObjectView data={translate(data)} />;
         } catch (err) {
             console.error("Couldn't parse data", err);
             Sentry.captureException(err);
@@ -128,7 +128,7 @@ const ObjectResult = (): JSX.Element => {
     const { state } = useLocation();
 
     if (instanceOfDataType(state)) {
-        return <ObjectLoaded data={state} />;
+        return <ObjectView data={state} />;
     }
 
     if (objID !== undefined) {
